+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Add support for PEP 742, adding `typing_extensions.TypeIs`. Patch
   by Jelle Zijlstra.
+- Drop runtime error when a mutable `TypedDict` key overrides a read-only
+  one. Type checkers should still flag this as an error. Patch by Jelle
+  Zijlstra.
 - Speedup `issubclass()` checks against simple runtime-checkable protocols by
   around 6% (backporting https://github.com/python/cpython/pull/112717, by Alex
   Waygood).